Renamed apio.proto to apio-scons.proto in preparation for additional apio proto files. No semantic change.
