+++ apio/commands/apio_graph.py
@@ -20,7 +20,11 @@ from apio.apio_context import (
     RemoteConfigPolicy,
 )
 from apio.utils import cmd_util
-from apio.common.proto.apio_pb2 import GraphOutputType, GraphParams, Verbosity
+from apio.common.proto.apio_scons_pb2 import (
+    GraphOutputType,
+    GraphParams,
+    Verbosity,
+)
 
 
 # ---------- apio graph
