+++ apio/scons/apio_env.py
@@ -17,7 +17,7 @@ import SCons.Defaults
 from apio.common.apio_console import cout
 from apio.common.apio_styles import EMPH3
 from apio.common.common_util import env_build_path
-from apio.common.proto.apio_pb2 import SconsParams
+from apio.common.proto.apio_scons_pb2 import SconsParams
 
 
 class ApioEnv:
