+++ apio/scons/plugin_base.py
@@ -25,7 +25,7 @@ from apio.common.apio_console import cout
 from apio.common.apio_styles import SUCCESS
 from apio.common.common_util import SRC_SUFFIXES
 from apio.scons.apio_env import ApioEnv
-from apio.common.proto.apio_pb2 import GraphOutputType
+from apio.common.proto.apio_scons_pb2 import GraphOutputType
 from apio.scons.plugin_util import (
     verilog_src_scanner,
     get_constraint_file,
